Add second email to the CoC
This adds my email to the Code of Conduct. Having two emails means that people may feel safer reporting on anyone in the community, as it gives them an option, especially in cases where the offender may have been me or @nshCore (hopefully, this never happens). It would be a good idea to set up community@catalystnet.org as an alias which can go to the community committee, and to add this email here when that is set up, as well.